% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_wts.R
\name{calculate_wts}
\alias{calculate_wts}
\title{Calculate weights that obey marginal balance constraints on protected characteristics.}
\usage{
calculate_wts(data, a = NULL, avals = NULL, wvars, epsilon = 1e-06)
}
\arguments{
\item{data}{data.frame containing the information to analyze}

\item{a}{optional string identifying the column in `data` that denotes the units of interest; if not provided, must provide `avals`}

\item{avals}{optional string identifying unique values of `a` found in the data; if not provided, must provide `a`}

\item{wvars}{string vector identifying the columns in `data` that denote the protected characteristics}

\item{epsilon}{positive scalar that indicates the amount that the optimization is allowed to deviate from the required constraints}
}
\description{
These weights ensure that each characteristic represented by a variable in `wvars` is balanced in the weighted population.
}
