% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqwt_plugin.R
\name{eqwt_plugin}
\alias{eqwt_plugin}
\title{Plug-in estimator for equity-weighted unit quality.}
\usage{
eqwt_plugin(
  data,
  folds = NULL,
  id,
  a,
  y,
  wvars,
  zvars,
  truncation_pt = NULL,
  adaptive = FALSE,
  K = 2,
  lrnr = lrn("classif.ranger"),
  lrnr_e = NULL,
  lrnr_mu = NULL,
  separate_e = FALSE,
  separate_mu = FALSE,
  epsilon = 1e-12,
  calibrate_e = TRUE,
  calibrate_mu = TRUE,
  condition_on = NULL
)
}
\arguments{
\item{data}{data.frame containing the information to analyze}

\item{folds}{optional string identifying the column in `data` that denotes the folds for cross-fitting}

\item{id}{string identifying the column in `data` that denotes the ID variable}

\item{a}{string identifying the column in `data` that denotes the units of interest}

\item{y}{string identifying the column in `data` that denotes the outcome of interest}

\item{wvars}{string vector identifying the columns in `data` that denote the protected characteristics}

\item{zvars}{string vector identifying the columns in `data` that denote the control (non-protected) characteristics}

\item{truncation_pt}{a number in (0,1) to which the propensity score is truncated (from above and below); default is 5/sqrt(n)/log(n)}

\item{adaptive}{logical flag of whether to use the adaptive normalization of Khan and Ugander (2021)}

\item{K}{optional integer identifying the number of folds for cross-fitting; required if folds = NULL}

\item{lrnr}{mlr3 learner object which will be used to estimate the mean function and propensity score}

\item{lrnr_e}{mlr3 learner object which will be used to estimate the propensity score (ignored if `lrnr` is specified)}

\item{lrnr_mu}{mlr3 learner object which will be used to estimate the mean function (ignored if `lrnr` is specified)}

\item{separate_e}{logical flag for whether propensity scores for each unit should be estimated separately or in a big multinomial model}

\item{separate_mu}{logical flag for whether mean functions for each unit should be estimated separately or in a big joint model}

\item{epsilon}{positive scalar that indicates the amount that the optimization of equity balance constraints is allowed to deviate from the required constraints}

\item{calibrate_e}{logical flag for whether propensity scores should be calibrated after fitting}

\item{calibrate_mu}{logical flag for whether mean functions should be calibrated after fitting}

\item{condition_on}{a string indicating a variable within which to estimate conditional quality estimates.}
}
\value{
A `tibble` with the following columns:\itemize{
  \item a column with the same name as the argument \code{a} which indicates the unit.
  \item \code{plugin_est}: the plug-in estimate.
  \item \code{plugin_se}: the estimated standard error for the plug-in.
  \item \code{shrinkage_est}: an estimate that is shrunk using empirical bayes via the \code{ebnm} package.
  \item \code{shrinkage_est_se}: the estimated standard error for the shrinkage estimate.
  \item \code{reliability}: the reliability of the shrinkage estimate.
}
}
\description{
Plug-in estimator for equity-weighted unit quality.
}
