% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_eqwt_tmle.R
\name{estimate_eqwt_tmle}
\alias{estimate_eqwt_tmle}
\title{TMLE estimator for unit quality.}
\usage{
estimate_eqwt_tmle(data, a, aval, y, w = NULL, e, mu, truncation_pt = 1e-12)
}
\arguments{
\item{data}{data.frame containing the information to analyze}

\item{a}{string identifying the column in `data` that denotes the units of interest}

\item{aval}{string identifying the particular unit being estimated}

\item{y}{string identifying the column in `data` that denotes the outcome of interest}

\item{w}{string identifying the column in `data` that denotes the equity weight}

\item{e}{string identifying the column in `data` that denotes the propensity score}

\item{mu}{string identifying the column in `data` that denotes the mean function}
}
\description{
TMLE estimator for unit quality.
}
