% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_unit_tmle.R
\name{estimate_unit_tmle}
\alias{estimate_unit_tmle}
\title{TMLE estimator for unit quality.}
\usage{
estimate_unit_tmle(
  data,
  a,
  aval,
  y,
  w = NULL,
  e,
  mu,
  truncation_pt = 1e-12,
  condition_on = NULL,
  separate = TRUE
)
}
\arguments{
\item{data}{data.frame containing the information to analyze}

\item{a}{string identifying the column in `data` that denotes the units of interest}

\item{aval}{string identifying the particular unit being estimated}

\item{y}{string identifying the column in `data` that denotes the outcome of interest}

\item{w}{string identifying the column in `data` that denotes the equity weight}

\item{e}{string identifying the column in `data` that denotes the propensity score}

\item{mu}{string identifying the column in `data` that denotes the mean function}

\item{condition_on}{a string indicating a variable within which to estimate conditional quality estimates.}

\item{separate}{logical indicating whether TMLE adjustment model should be fit only using data from the unit of interest}
}
\description{
TMLE estimator for unit quality.
}
