% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_wtd_plugin.R
\name{estimate_wtd_plugin}
\alias{estimate_wtd_plugin}
\title{Workhorse function for plug-in estimator. Intended to be wrapped by `popwt_plugin` and `eqwt_plugin`.}
\usage{
estimate_wtd_plugin(
  ds,
  folds,
  id,
  a,
  y,
  wvars,
  zvars,
  wt,
  truncation_pt = 1e-07,
  adaptive = FALSE,
  K = 2,
  lrnr = lrn("classif.ranger"),
  lrnr_e = NULL,
  lrnr_mu = NULL,
  separate_e = TRUE,
  separate_mu = TRUE,
  calibrate_e = FALSE,
  calibrate_mu = FALSE,
  infl_fn_only = FALSE,
  condition_on = NULL
)
}
\arguments{
\item{folds}{optional string identifying the column in `data` that denotes the folds for cross-fitting}

\item{id}{string identifying the column in `data` that denotes the ID variable}

\item{a}{string identifying the column in `data` that denotes the units of interest}

\item{y}{string identifying the column in `data` that denotes the outcome of interest}

\item{wvars}{string vector identifying the columns in `data` that denote the protected characteristics}

\item{zvars}{string vector identifying the columns in `data` that denote the control (non-protected) characteristics}

\item{wt}{string vector identifying the column in `data` that denotes the observation weights}

\item{truncation_pt}{a number in (0,1) to which the propensity score is truncated (from above and below)}

\item{adaptive}{logical flag of whether to use the adaptive normalization of Khan and Ugander (2021)}

\item{K}{optional integer identifying the number of folds for cross-fitting; required if folds = NULL}

\item{lrnr}{mlr3 learner object which will be used to estimate the mean function and propensity score}

\item{lrnr_e}{mlr3 learner object which will be used to estimate the propensity score (ignored if `lrnr` is specified)}

\item{lrnr_mu}{mlr3 learner object which will be used to estimate the mean function (ignored if `lrnr` is specified)}

\item{separate_e}{logical flag for whether propensity scores for each unit should be estimated separately or in a big multinomial model}

\item{separate_mu}{logical flag for whether mean functions for each unit should be estimated separately or in a big joint model}

\item{calibrate_e}{logical flag for whether propensity scores should be calibrated after fitting}

\item{calibrate_mu}{logical flag for whether mean functions should be calibrated after fitting}

\item{infl_fn_only}{logical flag for whether to return a tibble with the influence function for each observation and each unit (which can be used to calculate contrasts between units).}

\item{condition_on}{a string indicating a variable within which to estimate conditional quality estimates.}

\item{data}{data.frame containing the information to analyze}
}
\description{
Workhorse function for plug-in estimator. Intended to be wrapped by `popwt_plugin` and `eqwt_plugin`.
}
