% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_wtd_tmle.R
\name{estimate_wtd_tmle}
\alias{estimate_wtd_tmle}
\title{Workhorse function for TMLE. Intended to be wrapped by `popwt_tmle` and `eqwt_tmle`.}
\usage{
estimate_wtd_tmle(
  ds,
  folds,
  id,
  a,
  y,
  wvars,
  zvars,
  wt,
  truncation_pt = 1e-07,
  adaptive = FALSE,
  lrnr = lrn("classif.ranger"),
  lrnr_e = NULL,
  lrnr_mu = NULL,
  separate_e = TRUE,
  separate_mu = TRUE,
  calibrate_e = FALSE,
  calibrate_mu = FALSE,
  calibration_grps = 500,
  verbose = FALSE
)
}
\arguments{
\item{folds}{optional string identifying the column in `data` that denotes the folds for cross-fitting}

\item{id}{string identifying the column in `data` that denotes the ID variable}

\item{a}{string identifying the column in `data` that denotes the units of interest}

\item{y}{string identifying the column in `data` that denotes the outcome of interest}

\item{wvars}{string vector identifying the columns in `data` that denote the protected characteristics}

\item{zvars}{string vector identifying the columns in `data` that denote the control (non-protected) characteristics}

\item{wt}{string vector identifying the column in `data` that denotes the observation weights}

\item{truncation_pt}{a number in (0,1) to which the propensity score is truncated (from above and below)}

\item{lrnr}{mlr3 learner object which will be used to estimate the mean function and propensity score}

\item{lrnr_e}{mlr3 learner object which will be used to estimate the propensity score (ignored if `lrnr` is specified)}

\item{lrnr_mu}{mlr3 learner object which will be used to estimate the mean function (ignored if `lrnr` is specified)}

\item{separate_e}{logical flag for whether propensity scores for each unit should be estimated separately or in a big multinomial model}

\item{separate_mu}{logical flag for whether mean functions for each unit should be estimated separately or in a big joint model}

\item{calibrate_e}{logical flag for whether propensity scores should be calibrated after fitting}

\item{calibrate_mu}{logical flag for whether mean functions should be calibrated after fitting}

\item{calibration_grps}{tuning parameter for calibration, lower numbers induce more smoothness, with the default being 500}

\item{verbose}{logical flag for whether to print informative messages about progress.}

\item{data}{data.frame containing the information to analyze}
}
\description{
Workhorse function for TMLE. Intended to be wrapped by `popwt_tmle` and `eqwt_tmle`.
}
