% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_folds.R
\name{make_folds}
\alias{make_folds}
\title{Split dataset into folds.}
\usage{
make_folds(data, unit_id, K)
}
\arguments{
\item{data}{data.frame containing the information to analyze}

\item{unit_id}{string identifying the identifier for units in `data`}

\item{K}{number of folds}
}
\description{
Split dataset into folds.
}
