% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popwt_tmle.R
\name{popwt_tmle}
\alias{popwt_tmle}
\title{Targeted minimum loss-based estimation for unit quality using sample splitting.}
\usage{
popwt_tmle(
  data,
  folds = NULL,
  id,
  a,
  y,
  covars,
  truncation_pt = 1e-07,
  K = 2,
  lrnr = NULL,
  lrnr_e = NULL,
  lrnr_mu = NULL,
  separate_e = FALSE,
  separate_mu = FALSE,
  calibrate_e = TRUE,
  calibrate_mu = TRUE,
  calibration_grps = 500,
  condition_on = NULL,
  verbose = TRUE
)
}
\arguments{
\item{data}{data.frame containing the information to analyze}

\item{folds}{optional string identifying the column in `data` that denotes the folds for cross-fitting}

\item{id}{string identifying the column in `data` that denotes the ID variable}

\item{a}{string identifying the column in `data` that denotes the units of interest}

\item{y}{string identifying the column in `data` that denotes the outcome of interest}

\item{covars}{string vector identifying the columns in `data` that denote the control characteristics}

\item{truncation_pt}{a number in (0,1) to which the propensity score is truncated (from above and below); default is 5/sqrt(n)/log(n)}

\item{K}{optional integer identifying the number of folds for cross-fitting; required if folds = NULL}

\item{lrnr}{mlr3 learner object which will be used to estimate the mean function and propensity score}

\item{lrnr_e}{mlr3 learner object which will be used to estimate the propensity score (ignored if `lrnr` is specified)}

\item{lrnr_mu}{mlr3 learner object which will be used to estimate the mean function (ignored if `lrnr` is specified)}

\item{separate_e}{logical flag for whether propensity scores for each unit should be estimated separately or in a big multinomial model}

\item{separate_mu}{logical flag for whether mean functions for each unit should be estimated separately or in a big joint model}

\item{calibrate_e}{logical flag for whether propensity scores should be calibrated after fitting}

\item{calibrate_mu}{logical flag for whether mean functions should be calibrated after fitting}

\item{calibration_grps}{tuning parameter for calibration, lower numbers induce more smoothness, with the default being 500}

\item{condition_on}{a string indicating a variable within which to estimate conditional quality estimates.}

\item{verbose}{logical flag for whether to print informative messages about progress.}
}
\value{
A `tibble` with the following columns:\itemize{
  \item a column with the same name as the argument \code{a} which indicates the unit.
  \item \code{tmle_est}: the TMLE estimate.
  \item \code{tmle_se}: the estimated standard error for the TMLE.
  \item \code{shrinkage_est}: an estimate that is shrunk using empirical bayes via the \code{ebnm} package.
  \item \code{shrinkage_est_se}: the estimated standard error for the shrinkage estimate.
  \item \code{reliability}: the reliability of the shrinkage estimate.
}
}
\description{
This function is used for estimating many causal effects, all standardized to a common population distribution, with a primary application to quality measurement, where many units (e.g., doctors, hospitals, schools) are measured repeatedly (e.g., on different patients), and we want to know what the average outcome would be for each unit, if they were measured on the whole popuation (e.g., if each doctor saw all patients).
}
