% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrink_estimates.R
\name{shrink_estimates}
\alias{shrink_estimates}
\title{Shrink and obtain reliabilities for a given set of estimates of unit quality}
\usage{
shrink_estimates(
  data,
  ests,
  ses,
  est_nm = "shrinkage_est",
  rel_nm = "reliability",
  ...
)
}
\arguments{
\item{data}{data.frame containing the information to analyze}

\item{ests}{string identifying the column in `data` that denotes the estimates}

\item{ses}{string identifying the column in `data` that denotes the standard errors of the estimates}

\item{est_nm}{name for shrunken estimates in the returned data.frame}

\item{rel_nm}{name for reliability in the returned data.frame}
}
\description{
Shrink and obtain reliabilities for a given set of estimates of unit quality
}
