% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate quality data, used for the example in the vignette.}
\usage{
simulate_data(p, n, pi_x, mu_x, qq)
}
\arguments{
\item{p}{number of units (e.g., providers)}

\item{n}{vector of length `p` specifying number of per-unit measurements (e.g., number of patients per provider)}

\item{pi_x}{vector of length `p` specifying the unit-specific probability of binary covariate `X1`}

\item{mu_x}{vector of length `p` specifying the unit-specific mean of a normal (sd = 1) covariate `X2`}

\item{qq}{vector of length `p` specifying the true unit-specific quality}
}
\description{
Simulate quality data, used for the example in the vignette.
}
