% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_eqwt_tmle.R
\name{estimate_eqwt_tmle}
\alias{estimate_eqwt_tmle}
\title{TMLE estimator for unit quality.}
\usage{
estimate_eqwt_tmle(
  data,
  a,
  y,
  w = NULL,
  e,
  mu,
  truncation_pt,
  adaptive = FALSE
)
}
\arguments{
\item{data}{data.frame containing the information to analyze}

\item{a}{string identifying the column in `data` that denotes the binary variable indicating the unit of interest}

\item{y}{string identifying the column in `data` that denotes the outcome of interest}

\item{w}{string identifying the column in `data` that denotes the equity weight}

\item{e}{string identifying the column in `data` that denotes the propensity score}

\item{mu}{string identifying the column in `data` that denotes the mean function}

\item{truncation_pt}{a number in (0,1) to which the propensity score is truncated}

\item{adaptive}{logical flag of whether to use the adaptive normalization of Khan and Ugander (2021)}
}
\description{
TMLE estimator for unit quality.
}
