% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_wtd_tmle.R
\name{estimate_wtd_tmle}
\alias{estimate_wtd_tmle}
\title{Workhorse function for TMLE. Intended to be wrapped by `popwt_tmle` and `eqwt_tmle`.}
\usage{
estimate_wtd_tmle(
  ds,
  folds,
  id,
  a,
  y,
  wvars,
  zvars,
  wt,
  truncation_pt = 1e-07,
  adaptive = FALSE,
  K = 2,
  lrnr = lrn("classif.ranger"),
  separate_e = TRUE,
  separate_mu = TRUE,
  epsilon = 1e-12,
  tune = FALSE,
  callibrate_e = FALSE,
  callibrate_mu = FALSE,
  include_unweighted = TRUE
)
}
\arguments{
\item{folds}{optional string identifying the column in `data` that denotes the folds for cross-fitting}

\item{id}{string identifying the column in `data` that denotes the ID variable}

\item{a}{string identifying the column in `data` that denotes the units of interest}

\item{y}{string identifying the column in `data` that denotes the outcome of interest}

\item{wvars}{string vector identifying the columns in `data` that denote the protected characteristics}

\item{zvars}{string vector identifying the columns in `data` that denote the control (non-protected) characteristics}

\item{wt}{string vector identifying the column in `data` that denotes the observation weights}

\item{truncation_pt}{a number in (0,1) to which the propensity score is truncated (from above and below)}

\item{adaptive}{logical flag of whether to use the adaptive normalization of Khan and Ugander (2021)}

\item{K}{optional integer identifying the number of folds for cross-fitting; required if folds = NULL}

\item{lrnr}{mlr3 learner object which will be used to estimate the mean function and propensity score}

\item{separate_e}{logical flag for whether propensity scores for each unit should be estimated separately or in a big multinomial model}

\item{separate_mu}{logical flag for whether mean functions for each unit should be estimated separately or in a big joint model}

\item{epsilon}{positive scalar that indicates the amount that the optimization of equity balance constraints is allowed to deviate from the required constraints}

\item{tune}{logical flag for whether tuning should be performed on the learners before estimating nuisance functions}

\item{callibrate_e}{logical flag for whether propensity scores should be callibrated after fitting}

\item{callibrate_mu}{logical flag for whether mean functions should be callibrated after fitting}

\item{include_unweighted}{logical flag for whether to also estimate a non-equity-weighted version of the estimator}

\item{data}{data.frame containing the information to analyze}
}
\description{
Workhorse function for TMLE. Intended to be wrapped by `popwt_tmle` and `eqwt_tmle`.
}
