% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popwt_tmle.R
\name{popwt_tmle}
\alias{popwt_tmle}
\title{TMLE for unit quality.}
\usage{
popwt_tmle(
  data,
  folds = NULL,
  id,
  a,
  y,
  covars,
  truncation_pt = 1e-07,
  adaptive = FALSE,
  K = 2,
  lrnr = lrn("classif.ranger"),
  lrnr_e = NULL,
  lrnr_mu = NULL,
  separate_e = FALSE,
  separate_mu = FALSE,
  calibrate_e = TRUE,
  calibrate_mu = TRUE
)
}
\arguments{
\item{data}{data.frame containing the information to analyze}

\item{folds}{optional string identifying the column in `data` that denotes the folds for cross-fitting}

\item{id}{string identifying the column in `data` that denotes the ID variable}

\item{a}{string identifying the column in `data` that denotes the units of interest}

\item{y}{string identifying the column in `data` that denotes the outcome of interest}

\item{covars}{string vector identifying the columns in `data` that denote the control characteristics}

\item{truncation_pt}{a number in (0,1) to which the propensity score is truncated (from above and below); default is 5/sqrt(n)/log(n)}

\item{adaptive}{logical flag of whether to use the adaptive normalization of Khan and Ugander (2021)}

\item{K}{optional integer identifying the number of folds for cross-fitting; required if folds = NULL}

\item{lrnr}{mlr3 learner object which will be used to estimate the mean function and propensity score}

\item{lrnr_e}{mlr3 learner object which will be used to estimate the propensity score (ignored if `lrnr` is specified)}

\item{lrnr_mu}{mlr3 learner object which will be used to estimate the mean function (ignored if `lrnr` is specified)}

\item{separate_e}{logical flag for whether propensity scores for each unit should be estimated separately or in a big multinomial model}

\item{separate_mu}{logical flag for whether mean functions for each unit should be estimated separately or in a big joint model}

\item{calibrate_e}{logical flag for whether propensity scores should be calibrated after fitting}

\item{calibrate_mu}{logical flag for whether mean functions should be calibrated after fitting}
}
\description{
TMLE for unit quality.
}
